requirements.txt
python-dotenv
openai 0.28.1
assemblyai 0.20.0
flask
